# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface JarvisStatus {
  connected: boolean;
  providerConfigured: boolean;
  providerName: string;
  version: string;
  externalApisEnabled: boolean;
  /** @nullable */
  error: string | null;
}

export interface JarvisMessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  goal: string;
}

export interface JarvisExecutionStep {
  stepId: string;
  objective: string;
  tool: string;
  output: string;
  /** @nullable */
  error: string | null;
  verified: boolean;
  verification: string;
}

export interface JarvisMessageResponse {
  success: boolean;
  goal: string;
  response: string;
  providerName: string;
  demoMode: boolean;
  /** @nullable */
  demoLabel: string | null;
  /** @nullable */
  planGoal: string | null;
  /** @nullable */
  planProvider: string | null;
  executionSteps: JarvisExecutionStep[];
  /** @nullable */
  failure: string | null;
}

export interface JarvisError {
  error: string;
  code: string;
}

export interface JarvisComponentHealth {
  component: string;
  healthy: boolean;
  state: string;
  details: string;
}

export type JarvisNetworkStateConnectivity = typeof JarvisNetworkStateConnectivity[keyof typeof JarvisNetworkStateConnectivity];


export const JarvisNetworkStateConnectivity = {
  online: 'online',
  degraded: 'degraded',
  offline: 'offline',
  local_only: 'local_only',
  recovering: 'recovering',
} as const;

export interface JarvisNetworkState {
  connectivity: JarvisNetworkStateConnectivity;
  reachableHosts: string[];
  unreachableHosts: string[];
  details: string;
}

export interface JarvisRecoveryIncident {
  identifier: string;
  serviceName: string;
  reason: string;
  restartCount: number;
  timestamp: string;
  resolved: boolean;
}

export interface JarvisSecuritySummary {
  alertCount: number;
  findingCount: number;
  highestSeverity: string;
  /** @nullable */
  lastAssessmentAt: string | null;
}

export interface JarvisAgentActivity {
  taskId: string;
  agentName: string;
  success: boolean;
  summary: string;
  timestamp: string;
}

export interface JarvisSystemReport {
  demoMode: boolean;
  /** @nullable */
  demoLabel: string | null;
  health: JarvisComponentHealth[];
  network: JarvisNetworkState;
  recentIncidents: JarvisRecoveryIncident[];
  security: JarvisSecuritySummary;
  recentAgentActivity: JarvisAgentActivity[];
}