# lib/api-zod/src/generated/types/jarvisExecutionStep.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JarvisExecutionStep {
  stepId: string;
  objective: string;
  tool: string;
  output: string;
  /** @nullable */
  error: string | null;
  verified: boolean;
  verification: string;
}